refactor system monitor and add ima visibility
