+++ KubeArmor/BPF/common_types.h
@@ -1,0 +1,9 @@
+#ifndef _COMMON_TYPES_H
+#define _COMMON_TYPES_H
+
+struct outer_key {
+    u32 pid_ns;
+    u32 mnt_ns;
+};
+
+#endif // _COMMON_TYPES_H
