+++ KubeArmor/BPF/throttling.h
@@ -4,15 +4,12 @@
 #ifndef __THROTTLING_H
 #define __THROTTLING_H
 
+#include "common_types.h"
+
 struct alert_throttle_state {
   u64 first_event_timestamp;
   u64 event_count;
   u64 throttle;
-};
-
-struct outer_key {
-  u32 pid_ns;
-  u32 mnt_ns;
 };
 
 struct alert {
